move encoding out of core
